% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangeHTML.r
\name{arrangeHTML}
\alias{arrangeHTML}
\title{Arrange figures in HTML}
\usage{
arrangeHTML(x, title = "untitled")
}
\arguments{
\item{x}{A data frame listing the graphics files to embed. See details below.}

\item{title}{Character string used as a title for the HTML page.}
}
\value{
HTML code as a character string. Needs to be printed to a file, e.g.
  using \code{write}.
}
\description{
Embed a collection of figures in HTML with basic navigation capabilities
}
\note{
The data frame \code{x} must have 4 columns named 'h1', 'h2', 'width',
  and 'file' all being of character type. The first two columns specify
  headers of category 1 and 2, respectively. The 'h2' column can contain
  \code{NA} values where sub-headings are to be suppressed.
  The width is best specified as a percentage,
  e.g. '50%'. The last column contains the names of the graphics files to
  display. The files need to be in a format supported by the web browser
  such as SVG or PNG for vector and pixel graphics, respectively.

  The data frame is \emph{not} sorted by any column (not even 'h1').
}
\examples{
test= data.frame(stringsAsFactors=FALSE,
  h1=c("summer","summer","winter","winter"),
  h2=c("TP","TN","TP","TN"),
  width="50\%",
  file=c("summerTP.svg","summerTN.svg","winterTP.svg","winterTN.svg")
)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
