% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{coli}
\alias{coli}
\title{A small set of Escherichia coli growth curves (optical density over time)}
\description{
A small set of Escherichia coli growth curves (optical density over time)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
