% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinctSequences.r
\name{extinctSequences}
\alias{extinctSequences}
\title{Generates random sequences of forced extinctions}
\usage{
extinctSequences(p, r)
}
\arguments{
\item{p}{Named numeric vector of extinction probabilities. Names
represent species. The values should be in range 0, 1 but the do not have
to sum up to 1 (see notes on the \code{prob} argument of function 
\code{\link[base]{sample}}).}

\item{r}{The number of sequences to be generated. Should be 1 at least.}
}
\value{
A matrix with \code{r} columns and \code{length(p)} rows filled with
  species names sampled from the \code{names(p)}. Each column represents an
  individual orderer extinction sequence. Species with larger values of
  \code{p} are likely to occur at an earlier position in the sequence, i.e.
  closer to the top of the matrix.
}
\description{
Generates random sequences of forced extinctions. The likelihood of
  extinction for a particular species is controlled by a numeric vector of
  probabilities.
}
\examples{
# Basic example
p <- c("top"= 0.9, "fish1"= 0.8, "fish2"= 0.7, "zoopl"= 0.5, "algae"= 0.1)
print(extinctSequences(p, r=5))
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
