% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinctSimplify.r
\name{extinctSimplify}
\alias{extinctSimplify}
\title{Simplify a consumer-resource matrix}
\usage{
extinctSimplify(m, drop, keep)
}
\arguments{
\item{m}{Square matrix filled with 0 and 1. Column names represent consumers,
row names represent ressources. Row and column names must be identical.
A value of 1 indicates that the consumer in the respective column feeds on
the resource in the respective row.}

\item{drop}{Vector of character strings defining the rows and columns to be
removed from the food web (forced extinctions).}

\item{keep}{Vector of character strings defining the rows and columns to be
always kept in the system. Typically, these represent primary producers.
The elements in \code{keep} should not be present in \code{drop} as well.}
}
\value{
A square matrix with less rows and columns than the input matrix
  \code{m}. The result matrix reflects both the forced extinctions (specified
  in \code{drop}) and the secondary extinctions being triggered by the forced
  ones.
}
\description{
Computes the effect of forced extinctions on a consumer-resource matrix.
}
\examples{
# A simple aquatic food consisting of algae, zooplankton
# planktivorous fish and a top predator.

m <- rbind(
  algae =  c(0,1,1,0,0,0),
  zoopl1 = c(0,0,0,1,1,0),
  zoopl2 = c(0,0,0,0,1,0),
  fish1 =  c(0,0,0,0,0,1),
  fish2 =  c(0,0,0,0,0,1),
  top =    c(0,0,0,0,0,0)
)
colnames(m) <- rownames(m)
print("Original food web:")
print(m)

m2 <- extinctSimplify(m,
  drop=c("zoopl1","fish2"), keep=c("algae"))
print("Resulting food web:")
print(m2)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
