% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherPower.r
\name{fisherPower}
\alias{fisherPower}
\title{Power analysis for fisher test}
\usage{
fisherPower(pA, pB, nA, nB, alpha = 0.05, nRepl = 1000)
}
\arguments{
\item{pA}{True proportion in 1st sample. Not to be confused with odds.}

\item{pB}{True proportion in 2nd sample. Not to be confused with odds.}

\item{nA}{Size of 1st sample.}

\item{nB}{Size of 2nd sample.}

\item{alpha}{Alpha error (default 0.05).}

\item{nRepl}{Number of simulations to be performed.}
}
\value{
The power of the test, i.e. 1-Beta.
}
\description{
Computes the power of fishers exact test by means of simulation.
}
\examples{
nA <- 814
pA <- 29/814
nB <- 276
pB <- 1/276
print(fisherPower(pA, pB, nA, nB))
\dontrun{
library("statmod")
power.fisher.test(p1=pA, p2=pB, n1=nA, n2=nB, alpha=0.05, nsim=1000, alternative="two.sided")
}
}
\seealso{
Method \code{power.fisher.test} from package \code{statmod}
  yields essentially the same results.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
