% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.r
\name{lift}
\alias{lift}
\title{Linear fitting with interactive data selection}
\usage{
lift(df, time = "time", value = "value", keys = c("ID"), trans = log,
  ylim = NULL)
}
\arguments{
\item{df}{A data frame with columns specified in the other arguments.}

\item{time}{Name of the column of \code{df} holding numeric times
corresponding to observations. Should be a character string.}

\item{value}{Name of the column of \code{df} holding the observed
time dependent quantity, e.g. optical density. Should be a character string.}

\item{keys}{Name(s) of the column(s) of \code{df} that define subsets of the
data to be analyzed separatety. Should be a vector of character strings
which can be of length one. The specified columns should be of type
character, integer, or factor. They typically hold experiment IDs and/or
replicate IDs.}

\item{trans}{A function that is applied to the data in the \code{value}
column of \code{df}. The default, \code{\link[base]{log}}, is useful
for microbial growth data, for example.}

\item{ylim}{Either \code{NULL} or a numeric vector of length 2 specifying the
lower and upper limit for the value axis, respectively. Note that the
values must be supplied in the transformed scale if \code{trans} is used.}
}
\value{
A data frame with one row for each unique combination of the key
  columns of \code{df}. The number of columns is \code{length(keys)} plus 5.
  The contents of the additional columns is as follows:
  \itemize{
    \item{slope.default}: Slope of linear model fitted with \code{\link[stats]{lm}}
    \item{slope.robust}: Slope of linear model fitted with \code{\link[robust]{lmRob}}
    \item{used.points}: Number of points selected for fitting
    \item{time.first}: Time coordinate of first (earliest) point selected
    \item{time.last}: Time coordinate of last (latest) point selected
  }
}
\description{
The function fits linear models to a subset of data that is interactively
selected from a scatter plot.
}
\note{
A plot is created for each subset of the data and the user is asked to
  select a (time) range on the x-axis by two sebsequent clicks with the left
  mouse button. The selected data points are highlighted and the fitted
  models are displayed. The selection can be corrected by repeating the
  two subsequent left-clicks. A click with the right mouse button (or
  pressing the 'finish' button in Rstudio) accepts
  the current result and the process continues with the next subset of the
  data (i.e. the next plot). The results is returned after all subsets
  were processed. Since the function does not return intermediate output
  you better call it for smaller data sets (e.g. < 100 plots) only. This
  avoids the loss of too much manual work in case of a crash, blackout, ...
}
\examples{
\dontrun{
d <- read.table(header=TRUE, sep="", text='
  ID  time  value
  1   1     1
  1   2     10
  1   3     150
  1   4     900
  2   1     NA
  2   2     NA
  2   3     NA
  3   1     2
  3   2     33
  3   3     180
  3   4     3500
')
print(lift(d))
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
