% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPercentageMatrix.r
\name{plotColorMatrix}
\alias{plotColorMatrix}
\title{Plots a numeric matrix using a color scale}
\usage{
plotColorMatrix(x, breaks = c(50, 20, 10, 5, 2, 0)/100, colors = if
  (useHatching) c("black", "white") else c("orangered3", "khaki", "royalblue4"),
  useHatching = FALSE, translateRownames = function(x) {     x },
  translateColnames = function(x) {     x }, mar = c(5, 5, 1, 8),
  rowGroups = rep(1, nrow(x)))
}
\arguments{
\item{x}{Numeric matrix with column and row names.}

\item{breaks}{Numeric vector defining breaks in the legend. This typically
includes the lower limit (e.g. zero for percentage data) but not the
upper limit. The default is useful for fraction data (range 0...1).}

\item{colors}{A vector of color names to build the palette.}

\item{useHatching}{Use hatching to improve the distinction between
neighboring colors? (logical)}

\item{translateRownames}{Function to translate row names. Must accept a vector
and return a vector of character strings of the same length as the input}

\item{translateColnames}{Function to translate column names. Must accept a vector
and return a vector of character strings of the same length as the input}

\item{mar}{Vector of length 4 passed to the \code{mar} element of
\code{\link[graphics]{par}}. The last element (right margin) must be chosen
large enough so as to not cut off the automatic legend.}

\item{rowGroups}{A vector of e.g. factors splitting the rows of \code{x}
into groups. Lines will be plotted at the transition between groups.}
}
\value{
\code{invisible(NULL)}.
}
\description{
Plots a numeric matrix using a color scale.
}
\note{
Note that the matrix is plotted in transposed form.
}
\examples{
x <- matrix(runif(24), nrow=4, ncol=6,
 dimnames=list(paste("row", 1:4), paste("col", 1:6)))
plotColorMatrix(x=x, breaks=c(50,20,10,5,2,0)/100,
 useHatching=TRUE, translateColnames=function(x){sapply(x, gsub,
   patter="col", replacement="column", fixed=TRUE)}, rowGroups= c(1,1,2,2))
plotColorMatrix(x=x, breaks=c(50,20,10,5,2,0)/100,
 useHatching=FALSE, translateColnames=function(x){sapply(x, gsub,
 patter="col", replacement="column", fixed=TRUE)}, rowGroups= c(1,1,2,2))
}
\references{
No references.
}
\seealso{
No links.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
