% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGrowth.r
\name{prepGrowth}
\alias{prepGrowth}
\title{Prepares UHEL growth rate data for further analysis}
\usage{
prepGrowth(file_plate, file_growth, file_out, sep_in = "\\t", dec_in = ".",
  overwrite = FALSE)
}
\arguments{
\item{file_plate}{Delimited text file with info on plate layout. Must have
columns 'Well_ID', 'Strain', 'Replicate'.}

\item{file_growth}{Delimited text file with observed data. Must have column
'Time'. All additional columns must be named after wells as specified in
the field 'Well_ID' of the other input file or the colum must carry the
name 'blank'.
".}

\item{file_out}{Name for the output file.}

\item{sep_in}{The field delimiter for input files. Default is TAB and the
later is always used for the output.}

\item{dec_in}{The decimal character for input files. Default is period and
the latter is always used for the output.}

\item{overwrite}{Logical. Should existing output files be overwritten?}
}
\value{
Nothing. All results are writte to the specified output file.
}
\description{
Prepares growth rate data produced by Stefanie at UHEL for further analysis.
The function merges info from two separate files and also subtracts the
blank value.
}
\examples{
\dontrun{
  growthDB(file_plate="in/strains.txt", file_growth="in/growth.txt",
    file_out="out.txt")
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
