% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeoCompare.r
\name{rodeoCompare}
\alias{rodeoCompare}
\title{Simulate and compare with observations}
\usage{
rodeoCompare(inputs, model, obs, extract)
}
\arguments{
\item{inputs}{Named numeric vector. This holds the values of parameters and/or
initial values.}

\item{model}{Object of class \code{rodeo}.}

\item{obs}{Numeric matrix or data frame with numeric times in the first
column and observed data in the second column.}

\item{extract}{A function that extracts simulated data from the output matrix
of \code{simul} so that the result can compared with the second column of
\code{obs}.}
}
\value{
Numeric matrix with three colums named 'time', 'sim', and 'obs'.
}
\description{
Runs a model for specific parameter and initial values and arranges
times series of simulated and observed values in a data frame
(e.g. for plotting or calculation of residuals).
}
\note{
The function \code{extract} typically extracts a single column or merges
  the values from multiple columns using \code{apply}.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
