% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeoCompareAll.r
\name{rodeoCompareAll}
\alias{rodeoCompareAll}
\title{Simulate and compare with observations for multiple data sets}
\usage{
rodeoCompareAll(globals, locals, model, lst_obs, extract, t0 = NULL)
}
\arguments{
\item{globals}{Named numeric vector holding parameters and/or initial values.
These are global in the sense that they are common to all elements of
\code{obs}.}

\item{locals}{A numeric matrix with row and column names. Each row holds the
values of parameters and/or initial values applying to a specific
set of observations. The matching between \code{locals} and \code{lst_obs}
is deduced from the row and element names, respectively.}

\item{model}{Object of class \code{rodeo}.}

\item{lst_obs}{A list of numeric matrices, each of which holds numeric times in
the first column and values of observed variables in the remaining columns.}

\item{extract}{A function that extracts simulated data from the output matrix
of \code{simul}. It should return a matrix of dimensions \eqn{nt * nv}
where \eqn{nt} is the number of model output times and \eqn{nv} equals the
number of value columns in \code{obs}, i.e. excluding the time column.
Typically, the function selects a specific column (or multiple colums) from
the original output matrix of the ODE solver or it computes new column(s)
from the original ones (e.g. the sum of two state variables).}

\item{t0}{Time corresponding to the supplied initial values in argument
\code{inputs}. If \code{NULL}, the element \code{obs[1,1]} will be used.
The value typically needs to be set if the observations are delayed with
respect to the start of the simulation period.}
}
\value{
A data frame with five columns:
\itemize{
  \item{id}: Name of the set of observations
  \item{item}: Name of the observed / simulated entity
  \item{time}: Time
  \item{obs}: Observed value
  \item{sim}: Simulated value
}
}
\description{
Extends \code{rodeoCompare} for use with multiple data sets.
}
\note{
No notes.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
