% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeoEstimate.r
\name{rodeoEstimate}
\alias{rodeoEstimate}
\title{Estimate parameters}
\usage{
rodeoEstimate(inMod.guess, inMod.lower, inMod.upper, lst_inFix, model, lst_obs,
  extract, transf)
}
\arguments{
\item{inMod.guess}{Initial guess of the parameters / initial values to be
optimized. Named numeric vector.}

\item{inMod.lower}{Lower bounds of the parameters / initial values to be
optimized. Named numeric vector.}

\item{inMod.upper}{Upper bounds of the parameters / initial values to be
optimized. Named numeric vector.}

\item{lst_inFix}{A list of named numeric vectors. Each vector holds the values
of parameters and/or initial values being constant for the corresponding
set of observations. The length of the list must match the length of the
list \code{obs}.}

\item{model}{Object of class \code{rodeo}.}

\item{lst_obs}{A list of 2-column matrices (see argument \code{obs} of \code{compare}).
Each list element specifies an independent set of observations.}

\item{extract}{A function that extracts simulated data from the output matrix
of \code{simul} so that the result can compared with the second column of
\code{obs}.}

\item{transf}{A function being applied to the observed and simulated data
before computing residuals. Its single argument should be a numeric vector.
The default function doesn't perform any transformation. A typical example
is \code{log10}.}
}
\value{
A list with the following components:
\itemize{
  \item{estimates} Named vector of the estimated parameters.
  \item{stdErrors} Named vector of the parameters' standard errors.
  \item{residuals} Numeric vector of residuals (e.g. for calculation of RMSE).
  \item{info} Return code of optimizer. See help for \code{summary.modFit}.
}
}
\description{
Estimates parameters using \code{modFit} from the \code{FME} package.
}
\note{
See notes for \code{compare}.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
