% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeoInitialize.r
\name{rodeoInitialize}
\alias{rodeoInitialize}
\title{Set-up a compiled rodeo based model from text files}
\usage{
rodeoInitialize(dir)
}
\arguments{
\item{dir}{Directory holding text files with fixed names "vars.txt",
"pars,txt", "funs.txt", "pros.txt", "stoi.txt" as well as a
 Fortran source file "functions.f95".}
}
\value{
A rodeo object ready for compiled-code simulation.
}
\description{
Creates a rodeo object and compiles Fortran source code
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
