% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeoResiduals.r
\name{rodeoResiduals}
\alias{rodeoResiduals}
\title{Computes residuals between simulated and observed values}
\usage{
rodeoResiduals(inMod, lst_inFix, model, lst_obs, extract, transf)
}
\arguments{
\item{inMod}{Named numeric vector holding parameters and/or initial values.
These are global in the sense that they are common to all elements of
\code{obs}.}

\item{lst_inFix}{A list of named numeric vectors. Each vector holds the values
of parameters and/or initial values being constant for the corresponding
set of observations. The length of the list must match the length of the
list \code{obs}.}

\item{model}{Object of class \code{rodeo}.}

\item{lst_obs}{A list of 2-column matrices (see argument \code{obs} of \code{compare}).
Each list element specifies an independent set of observations.}

\item{extract}{A function that extracts simulated data from the output matrix
of \code{simul} so that the result can compared with the second column of
\code{obs}.}

\item{transf}{A function being applied to the observed and simulated data
before computing residuals. Its single argument should be a numeric vector.
The default function doesn't perform any transformation. A typical example
is \code{log10}.}
}
\value{
Numeric vector of residuals. The length of the return vector equals
  the total number of observations, i.e. \code{sum(sapply(lst_obs, nrow))}.
}
\description{
Computes residuals. This function is to be passed as argument \code{f} to
\code{modFit} from the \code{FME} package.
}
\note{
See notes for \code{compare}.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
