% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeoSimulate.r
\name{rodeoSimulate}
\alias{rodeoSimulate}
\title{Run simulation}
\usage{
rodeoSimulate(inputs, model, times)
}
\arguments{
\item{inputs}{Named numeric vector. This holds the values of parameters and/or
initial values.}

\item{model}{Object of class \code{rodeo}.}

\item{times}{Numeric vector specifying the desired output times.}
}
\value{
Numeric matrix with times in the first column and states or
  process rates in the remaining columns.
}
\description{
Runs a simulation of a 0-dimensional \code{rodeo}-based model.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
