% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.reformat.r
\name{plate.reformat}
\alias{plate.reformat}
\title{Prepares plate reader data for further analysis}
\usage{
plate.reformat(layout, dynamics)
}
\arguments{
\item{layout}{Data frame with info on plate layout. Must have
columns 'id_well', 'id_sample', 'id_replicate'. It is necessary that, for
one or more records, the value in column 'id_sample' is set to 'blank'.}

\item{dynamics}{Data frame with observed data. Must have column
'time'. All additional columns must be named after wells as specified in
the field 'id_well' of the other input file.
".}
}
\value{
A data frame.
}
\description{
Plate reader data are reformatted to facilitate further analysis.
Specifically, the function merges info from two separate files and also
subtracts the blank value.
}
\note{
The subtracted blank value is the mean of the blank wells for the
  respective time step. Possible negative results are set to zero.
}
\examples{
\dontrun{
  # Example to be prepared
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
