% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeoEstimateDefunct.r
\name{rodeoEstimate}
\alias{rodeoEstimate}
\title{Estimate parameters}
\usage{
rodeoEstimate(inMod.guess, inMod.lower, inMod.upper, lst_inFix, model, lst_obs,
  extract, transf)
}
\arguments{
\item{inMod.guess}{Initial guess of the parameters / initial values to be
optimized. Named numeric vector.}

\item{inMod.lower}{Lower bounds of the parameters / initial values to be
optimized. Named numeric vector.}

\item{inMod.upper}{Upper bounds of the parameters / initial values to be
optimized. Named numeric vector.}

\item{model}{Object of class \code{rodeo}.}

\item{extract}{A function that extracts simulated data from the output matrix
of \code{simul}. It should return a matrix of dimensions \eqn{nt * nv}
where \eqn{nt} is the number of model output times and \eqn{nv} equals the
number of value columns in \code{obs}, i.e. excluding the time column.
Typically, the function selects a specific column (or multiple colums) from
the original output matrix of the ODE solver or it computes new column(s)
from the original ones (e.g. the sum of two state variables).}
}
\value{
A list with the following components:
\itemize{
  \item{estimates} Named vector of the estimated parameters.
  \item{stdErrors} Named vector of the parameters' standard errors.
  \item{residuals} Numeric vector of residuals (e.g. for calculation of RMSE).
  \item{info} Return code of optimizer. See help for \code{summary.modFit}.
}
}
\description{
Estimates parameters using \code{modFit} from the \code{FME} package.
}
\note{
See notes for \code{compare}.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
